requirements.txt
asgiref==3.5.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dictdiffer==0.9.0
Django==3.2
django-ckeditor==6.5.1
django-ckeditor-5==0.2.0
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-model-utils==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-extensions==0.7.1
drf-spectacular==0.26.4
drf-yasg==1.21.5
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpath-rw==1.4.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.2
mongoengine==0.27.0
mongoengine-migrate==0.0.3
numpy==1.24.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pdfkit==1.0.0
Pillow==9.2.0
ply==3.11
psycopg2-binary==2.9.5
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0.1
referencing==0.30.0
requests==2.28.2
rpds-py==0.9.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
sqlparse==0.2.4
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.0.9